requirements.txt
streamlit
yfinance
beautifulsoup4
requests
plotly
scikit-learn
joblib
google-generativeai
python-dotenv
